Update useSelectedDateEvents to own loading, and utilize useRef instead of global dict.
