+++ sms_frontend/src/components/Events/EventCard.tsx
@@ -1,5 +1,5 @@
 import dayjs from "dayjs";
-import { useContext, useEffect, useState } from "react";
+import { useContext, useState } from "react";
 
 import {
   Delete,
